invoice: refactor mod into directory
